Prevent unnecessary string scanning.
The internal strcat_s() function already calls strlen() on the
source/destination strings to check for overflows. Instead of calling
into strcat() and scanning over the string again, we can simply call
memcpy() to copy the text over to the right place.

Apply the same optimization to strcpy_s(). memcpy() is almost always
faster than strcpy(), for the reason that testing against a counter is
faster than searching for a null byte in the input sequence.

This change allows the code to build with Nuxi CloudABI's C library[1].
This C library does not provide implementations for strcpy(), strcat()
and strncat(), as in addition to being potentially insecure, the
observation is that if you want to use these securely, you can almost
always use memcpy() instead.

[1] cloudlibc: https://github.com/NuxiNL/cloudlibc