Add missing include to config.hpp to primary_transform.hpp
This header declares a namespace with the BOOST_REGEX_DETAIL_NS
token from the config.hpp header, but never includes this header
unlike the other headers in this directory.
This means that if a user includes this header first, suddenly
all declarations are inside the literal 'BOOST_REGEX_DETAIL_NS'
namespace which breaks all code using this header.